Fixes late joins
JoinLate would give a turf as it's location, which would be immediately useless as it  points to space because the ship gets moved to another z level. JoinLate landmarks now designate chairs like how arrival shuttles do, allowing you to choose unique places for people to spawn. Changed job code slightly to default to the entry hall area instead of the arrivals shuttle.